+++ lib/reputation_system/models/evaluation.rb
@@ -32,7 +32,7 @@ module ReputationSystem
     validate :source_must_be_defined_for_reputation_in_network
 
     def self.find_by_reputation_name_and_source_and_target(reputation_name, source, target)
-      source_type = get_source_type_for_sti(source, target.class.name, reputation_name)
+      source_type = get_source_type_for_sti(source.class.name, target.class.name, reputation_name)
       ReputationSystem::Evaluation.find(:first,
                         :conditions => {:reputation_name => reputation_name.to_s,
                                         :source_id => source.id,
@@ -48,11 +48,23 @@ module ReputationSystem
                            :target_id => target.id, :target_type => target.class.name)
     end
 
+    # Override exists? class method.
+    class << self
+      alias :original_exists? :exists?
+
+      def exists?(options={})
+        if options[:source_type] && options[:target_type] && options[:reputation_name]
+          options[:source_type] = get_source_type_for_sti(options[:source_type], options[:target_type], options[:reputation_name])
+        end
+        original_exists? options
+      end
+    end
+
     protected
 
-      def self.get_source_type_for_sti(source, target_type, reputation_name)
+      def self.get_source_type_for_sti(source_type, target_type, reputation_name)
         valid_source_type = ReputationSystem::Network.get_reputation_def(target_type, reputation_name)[:source].to_s.camelize
-        source_class = source.class
+        source_class = source_type.constantize
         while source_class && valid_source_type != source_class.name && source_class.name != "ActiveRecord::Base"
           source_class = source_class.superclass
         end
@@ -60,7 +72,7 @@ module ReputationSystem
       end
 
       def set_source_type_for_sti
-        sti_source_type = self.class.get_source_type_for_sti(source, target_type, reputation_name)
+        sti_source_type = self.class.get_source_type_for_sti(source_type, target_type, reputation_name)
         self.source_type = sti_source_type if sti_source_type
       end